requirements.txt
Django>=4.2,<4.3
psycopg2-binary>=2.9.5
requests>=2.28
django-extensions>=3.2.0
python-dotenv>=0.21.0
gunicorn>=20.1.0
coverage>=6.5
pytest-django>=4.5
Werkzeug>=2.2
